requirements.txt
numpy>=1.24.0
mysql-connector-python>=8.0.0 